Tidy up
Haven't yet merged with Danny's repo

+++ openCLImageLoad/openCLUtilities.cpp
@@ -199,8 +199,8 @@ char *load_program_source(const char *filename)
         return 0;
 	
     stat(filename, &statbuf);
-    source = (char *) malloc(statbuf.st_size + 1);
-    fread(source, statbuf.st_size, 1, fh);
+    source = (char *) malloc((unsigned long)statbuf.st_size + 1);
+    fread(source, (unsigned long)statbuf.st_size, 1, fh);
     source[statbuf.st_size] = '\0';
 	
     return source;
